refactor(deleteFirebaseFile): add user id validation for firebase file deletion

+++ api/server/services/Files/Firebase/crud.js
@@ -161,6 +161,9 @@ function extractFirebaseFilePath(urlString) {
  */
 const deleteFirebaseFile = async (req, file) => {
   const fileName = extractFirebaseFilePath(file.filepath);
+  if (!fileName.includes(req.user.id)) {
+    throw new Error('Invalid file path');
+  }
   await deleteFile('', fileName);
 };
